This code is efficient. From now, a lot of things will change because of the new tasks we have to implement. This code is now a reference, it works. After this commit, it will evolve a lot

+++ test2.2.cpp
@@ -6,7 +6,7 @@ using namespace std;
 int main()
 {	
 	Network network(2);
-	int t(t_start), t_a(0), t_b(0);
+	double t(t_start), t_a(0), t_b(0);
 	double Icurrent(0.0);
 	
 	do {
